Containerization/requirements.txt
APScheduler==3.7.0
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-SocketIO==5.1.0
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
six==1.16.0
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
